refactor email template
